% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{add}
\alias{add}
\title{Add three numbers}
\usage{
add(x, y, z)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{z}{A number.}
}
\value{
The sum x + y + z.
}
\description{
Add three numbers
}
\examples{
add(1, 1, 1)
}
