% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic2-package.R
\docType{package}
\name{arithmetic2}
\alias{arithmetic2}
\alias{arithmetic2-package}
\title{\code{arithmetic2} package}
\description{
An example package that does basic arithmetic and relies on another package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertdj/arithmetic2}
}

}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{git@dahl-jacobsen.dk}

}
